- m_eob.c:ms_eob(): use proper conversion specifier
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@7339 82007160-df01-0410-b94d-b575c5fd34c7

+++ modules/m_eob.c
@@ -63,9 +63,9 @@ ms_eob(struct Client *source_p, int parc, char *parv[])
 
   if (MyConnect(source_p))
     sendto_realops_flags(UMODE_SERVNOTICE, L_ALL, SEND_NOTICE,
-                         "End of burst from %s (%u seconds)",
+                         "End of burst from %s (%ju seconds)",
                          source_p->name,
-                         (unsigned int)(CurrentTime - source_p->connection->firsttime));
+                         (CurrentTime - source_p->connection->firsttime));
 
   server_eob(source_p);
